Add using_dng_exif configuration option
This option allows users to choose whether to use some EXIF information
from the DNG files or not.

Add Windows properties options

This commit adds the ability to allows users to write Windows properties

Add run.bat for Windows platform to execute in venv

This commit adds a batch file that can be used to run the program
in a virtual environment on Windows.